+++ src/cuda-sim/ptx_parser.h
@@ -142,8 +142,8 @@ class ptx_recognizer {
   void add_literal_int(int value);
   void add_literal_float(float value);
   void add_literal_double(double value);
-  void add_address_operand(const char *identifier, int offset);
-  void add_address_operand2(int offset);
+  void add_address_operand(const char *identifier, long long offset);
+  void add_address_operand2(long long offset);
   void add_label(const char *idenfiier);
   void add_vector_spec(int spec);
   void add_space_spec(enum _memory_space_t spec, int value);
